# packages/api-client/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Shellwright control plane",
    "description": "Organisations, apps, and immutable configuration versions.\n\nErrors are RFC 9457 problem documents. The `type` URI and the `code` extension are stable; the `title` and `detail` are for people and may be reworded.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/orgs/{orgId}/assets": {
      "post": {
        "tags": ["Assets"],
        "summary": "Upload an image and get back an asset reference.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetResponse": {
        "required": ["reference", "contentType", "bytes", "width", "height", "hasAlpha", "deduplicated"],
        "type": "object",
        "properties": {
          "reference": {
            "type": "string",
            "description": "The `asset://sha256-…` reference to put in a configuration."
          },
          "contentType": {
            "type": "string",
            "description": "Media type, determined from the bytes."
          },
          "bytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "description": "Size.",
            "format": "int64"
          },
          "width": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "description": "Pixel width.",
            "format": "int32"
          },
          "height": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "description": "Pixel height.",
            "format": "int32"
          },
          "hasAlpha": {
            "type": "boolean",
            "description": "Whether any pixel is not fully opaque."
          },
          "deduplicated": {
            "type": "boolean",
            "description": "True when these exact bytes were already stored."
          }
        },
        "description": "An uploaded asset as the API reports it."
      },
...
    }
  },
...
}